+++ bench/bench-hopscotchmap.cpp
@@ -25,7 +25,7 @@ using namespace std;
 /********************************************************************************************/
 
 static void
-test_dict2(unsigned long  n)
+test_dict2(size_t  n)
 {
   tsl::hopscotch_map<unsigned long, unsigned long> dict;
 
@@ -63,7 +63,7 @@ namespace std {
 }
 
 static void
-test_dict_big(unsigned long  n)
+test_dict_big(size_t  n)
 {
   hopscotch_map<char_array_s, char_array_s> dict;
 
@@ -87,7 +87,7 @@ test_dict_big(unsigned long  n)
 
 /********************************************************************************************/
 static void
-test_dict_str(unsigned long  n)
+test_dict_str(size_t  n)
 {
   hopscotch_map<string, string> dict;
